fix: always remember cursor
Every time the selector table is updated, store the current selected
item in the table of last remembered cursors. This allows us to always
update the cursor from that table when reloading the tables' contents.
This way, the bug is fixed where the left table's cursor wouldn't update
when focusing an item after an undo/redo.